+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Now sort the toml file before writing it to disk.
+
 ## [0.1.0]
 
 - initial release