+++ CHANGELOG.md
@@ -5,10 +5,15 @@
 ### Added
 
 - Added remove-etag command under repo
+- Added message on add, clone, remove.
 
 ### Changed
 
 - Sort the toml file before writing.
+
+### Fixed
+
+- Fixed default shell config for non-windows.
 
 ## [0.1.0]